Testing pull request

+++ calc.py
@@ -2,7 +2,8 @@
 # '1 + 2' -> 3, or '1 - 2' -> -1.
 def compute(expression):
     num0, operator, num1 = expression.split(' ')
-    num0, num1 = float(num0), int(num1)
+    num0, num1 = float(num0), float(num1)
+
     if operator == '+':
         return num0 + num1
     elif operator == '-':
@@ -17,3 +18,5 @@ def compute(expression):
     
     
     # comments
+
+# test pull request
